fix: full support of rte for announcement bar
