+++ apps/web/app/components/blocks/TextCard/TextCardForm.vue
@@ -9,77 +9,7 @@
       <h2>{{ getEditorTranslation('text-group-label') }}</h2>
     </template>
 
-    <div v-if="runtimeConfig.enableRichTextEditorV2" data-testid="text-card-form-v2">
-      <EditorOptionsTabs
-        :model-value="editorMode"
-        test-id-prefix="mode"
-        :legend="getEditorTranslation('content-label')"
-        :options="editorModeOptions"
-        @update:model-value="editorMode = $event"
-      />
-
-      <div v-if="editorMode === 'wysiwyg'" class="py-2">
-        <EditorRichTextEditor
-          ref="contentRichTextEditor"
-          v-model="contentModel"
-          v-model:expanded="expandedToolbars.content"
-          :min-height="232"
-          :expandable="true"
-          :text-align="textCardBlock.text.textAlignment"
-          data-testid="rte-content"
-          @request-html-modal="handleRequestHtmlModal"
-        />
-      </div>
-
-      <div v-else class="py-2">
-        <div class="flex items-center justify-between">
-          <UiFormLabel for="html-editor" class="m-0">
-            {{ getEditorTranslation('html-editor-label') }}
-          </UiFormLabel>
-
-          <EditorRichTextEditorMenuButton
-            aria-label="Open HTML editor in fullscreen"
-            icon-name="fullscreen"
-            class="ml-2"
-            @click="toggleModal"
-          />
-        </div>
-        <SfTextarea
-          id="html-editor"
-          v-model="htmlDraft"
-          data-testid="html-editor"
-          rows="10"
-          class="min-h-[232px] mt-1 block w-full border rounded-md shadow-sm sm:text-sm font-mono"
-          :class="htmlErrors.length ? 'border-red-400' : 'border-gray-300'"
-          :aria-invalid="htmlErrors.length ? 'true' : 'false'"
-          :aria-describedby="ariaDescribedBy"
-        />
-
-        <div
-          v-if="htmlErrors.length"
-          id="html-editor-errors"
-          data-testid="html-editor-errors"
-          class="mt-2 rounded-md border border-red-300 bg-red-50 p-3 text-sm text-red-800"
-          role="alert"
-        >
-          <div class="font-semibold">{{ getEditorTranslation('html-invalid-label') }}</div>
-          <ul class="list-disc ml-5 mt-1">
-            <li v-for="(e, idx) in htmlErrors.slice(0, 3)" :key="idx">{{ e }}</li>
-          </ul>
-        </div>
-      </div>
-
-      <EditorHtmlEditor
-        v-if="modalOpen"
-        v-model="htmlDraft"
-        :aria-describedby="ariaDescribedBy"
-        :html-errors="htmlErrors"
-        @switch-to-wysiwyg="handleSwitchToWysiwygFromModal"
-        @close="toggleModal"
-      />
-    </div>
-
-    <div v-else data-testid="text-card-form">
+    <EditorRichTextEditorForm v-model="contentModel" :text-align="textCardBlock.text.textAlignment">
       <div class="py-2">
         <div class="flex justify-between mb-2">
           <UiFormLabel>{{ getEditorTranslation('pretitle-label') }}</UiFormLabel>
@@ -201,7 +131,7 @@
           </div>
         </div>
       </fieldset>
-    </div>
+    </EditorRichTextEditorForm>
   </UiAccordionItem>
 
   <UiAccordionItem
@@ -382,16 +312,6 @@ import {
 import type { TextCardFormProps, TextCardContent } from './types';
 const props = defineProps<TextCardFormProps>();
 
-const runtimeConfig = useRuntimeConfig().public;
-const modalOpen = ref(false);
-const toggleModal = () => {
-  modalOpen.value = !modalOpen.value;
-};
-
-const contentRichTextEditor = ref<{
-  openModal: () => void;
-} | null>(null);
-
 const route = useRoute();
 const { data } = useBlockTemplates(
   route?.meta?.identifier as string,
@@ -401,10 +321,6 @@ const { data } = useBlockTemplates(
 
 const { blockUuid } = useSiteConfiguration();
 const { findOrDeleteBlockByUuid } = useBlockManager();
-
-const expandedToolbars = ref({
-  content: true,
-});
 
 const textCardBlock = computed<TextCardContent>(() => {
   const rawContent = findOrDeleteBlockByUuid(data.value, props.uuid || blockUuid.value)?.content ?? {};
@@ -440,33 +356,6 @@ const contentModel = computed<string>({
     textCardBlock.value.text.htmlDescription = val ?? '';
   },
 });
-const editorModeOptions = computed(
-  (): Array<{ value: EditorMode; label: string; testId: string }> => [
-    { value: 'wysiwyg', label: getEditorTranslation('wysiwyg-label'), testId: 'mode-wysiwyg' },
-    { value: 'html', label: getEditorTranslation('html-label'), testId: 'mode-html' },
-  ],
-);
-
-const { editorMode, htmlDraft, htmlErrors, ariaDescribedBy, switchToHtmlMode, switchToWysiwygMode } = useHtmlEditorMode(
-  contentModel,
-  {
-    defaultMode: 'wysiwyg',
-    commitOnValid: true,
-    maxErrors: 5,
-  },
-);
-
-const handleRequestHtmlModal = () => {
-  switchToHtmlMode();
-  if (!modalOpen.value) toggleModal();
-};
-
-const handleSwitchToWysiwygFromModal = async () => {
-  if (modalOpen.value) toggleModal();
-  switchToWysiwygMode();
-  await nextTick();
-  contentRichTextEditor.value?.openModal();
-};
 
 const { isFullWidth } = useFullWidthToggleForContent(textCardBlock);
 
@@ -507,12 +396,7 @@ watch([isTransparent, backgroundColor], () => {
     "background-color-label": "Background Color",
     "padding-label": "Padding",
     "spacing-around": "Spacing around the text elements",
-    "keep-transparent-label": "Keep background transparent",
-    "content-label": "Content",
-    "wysiwyg-label": "Rich Text",
-    "html-label": "HTML",
-    "html-editor-label": "HTML editor",
-    "html-invalid-label": "HTML syntax issues"
+    "keep-transparent-label": "Keep background transparent"
   },
   "de": {
     "text-group-label": "Text",
@@ -537,12 +421,7 @@ watch([isTransparent, backgroundColor], () => {
     "background-color-label": "Background Color",
     "padding-label": "Padding",
     "spacing-around": "Spacing around the text elements",
-    "keep-transparent-label": "or keep transparent",
-    "content-label": "Content",
-    "wysiwyg-label": "Rich Text",
-    "html-label": "HTML",
-    "html-editor-label": "HTML editor",
-    "html-invalid-label": "HTML syntax issues"
+    "keep-transparent-label": "or keep transparent"
   }
 }
 </i18n>
